thanks Boss!
i want to put it on record that im only alive because of
the Creator.

between the government (both the incoming and outgoing
administrations) and main theres no way in the world i
should be alive.

i know 2025 is going to better, the Boss has made it
abundantly clear we all of those that want to harm me amd
DO harm others will be addressed.

thanks again the Creator.

lets do it!
-nameless